+++ app/src/main/java/com/olamide/findartt/activity/ArtworkActivity.java
@@ -50,7 +50,7 @@
 //import com.olamide.findartt.fragment.BuyFragment;
 //import com.olamide.findartt.models.Artwork;
 //import com.olamide.findartt.models.ArtworkSummary;
-//import com.olamide.findartt.models.FindArttResponse;
+//import com.olamide.findartt.models.api.FindArttResponse;
 //import com.olamide.findartt.models.User;
 //import com.olamide.findartt.utils.AppExecutors;
 //import com.olamide.findartt.utils.Converters;
