+++ app/src/main/java/com/olamide/findartt/fragment/BuyFragment.java
@@ -15,7 +15,7 @@
 //import com.olamide.findartt.R;
 //import com.olamide.findartt.models.ArtworkSummary;
 //import com.olamide.findartt.models.Buy;
-//import com.olamide.findartt.models.FindArttResponse;
+//import com.olamide.findartt.models.api.FindArttResponse;
 //import com.olamide.findartt.models.User;
 //import com.olamide.findartt.utils.ErrorUtils;
 //import com.olamide.findartt.utils.TempStorageUtils;
