+++ app/src/main/java/com/olamide/findartt/widget/arttViewsService.java
@@ -9,7 +9,7 @@
 //
 //import com.olamide.findartt.R;
 //import com.olamide.findartt.models.Artwork;
-//import com.olamide.findartt.models.FindArttResponse;
+//import com.olamide.findartt.models.api.FindArttResponse;
 //import com.olamide.findartt.models.User;
 //import com.olamide.findartt.utils.TempStorageUtils;
 //import com.olamide.findartt.utils.network.FindArttRepository;
